% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{beta_fmt}
\alias{beta_fmt}
\alias{ci_beta_fmt}
\alias{RR_fmt}
\alias{ci_RR_fmt}
\alias{p_fmt}
\alias{prop_fmt}
\alias{sd_fmt}
\alias{mean_fmt}
\alias{n_fmt}
\alias{median_fmt}
\alias{iqr_fmt}
\alias{range_fmt}
\title{Formatting functions}
\usage{
beta_fmt(b)

ci_beta_fmt(lwr, upr, sep = " to ")

RR_fmt(RR)

ci_RR_fmt(lwr, upr, sep = " to ")

p_fmt(p, digits = 3, eps = 10^-digits)

prop_fmt(x)

sd_fmt(x)

mean_fmt(x)

n_fmt(x)

median_fmt(x)

iqr_fmt(lwr, upr, sep = " to ", bracket = c("[", "]"))

range_fmt(lwr, upr, sep = " to ", bracket = c("[", "]"))
}
\description{
These are functions passed to either summary measure functions or fmt.glm_table. The idea is for you to be able to create your own modifications of them and pass them to the relevant functions, in order to fully customize the table.
}
