% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggroc.R
\name{ggroc}
\alias{ggroc}
\title{ggplot2 for ROC curves}
\usage{
ggroc(list_rocs, fmt = fmt_auc, labs)
}
\arguments{
\item{list_rocs}{A list of objects returned by \code{pROC::roc(ci=TRUE)}.}

\item{fmt}{A function receiving the parameters \code{roc} and \code{digits}, to return the AUC text in the annotation.}

\item{labs}{Character vector, same length as list_rocs, with labels for the different models.}

\item{roc}{An object returned by \code{pROC::roc(ci=TRUE)}.}

\item{digits}{Integer vector of length 1. The number of digits to print for AUC and 95% confidence interval.}
}
\description{
Plot one or more ROC curves using the ggplot2 interface.
}
