% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_table.R
\name{glm_table}
\alias{glm_table}
\title{Produce a formatted generalized linear model table}
\usage{
glm_table(fit, digits = 4, p.digits = digits + 1, se = "normal",
  intercept = FALSE, fmt = FALSE, fun_coef = if (fit$family$link \%in\%
  c("log", "logit")) exp else I, ...)
}
\arguments{
\item{fit}{An glm fit object of class \code{glm}.}

\item{digits}{Integer. The number of significant digits to return in coefficients}

\item{p.digits}{Integer. The numeric of significant digits to return in p-values}

\item{se}{Character. The type of standard error to use, "normal" or "robust".}

\item{intercept}{Logical. Whether or not to include the intercept row in the returned table.}

\item{fmt}{Logical. If true, passes the returned table to \code{fmt.glm_table}, returning a publication-ready table.}

\item{fun_coef}{Function to apply to the coefficients. By default, \code{exp} if link is "log" or "logit", otherwise identity.}

\item{...}{Arguments passed to \code{fmt.glm_table}.}
}
\value{
A matrix with coefficients, 95\% confidence intervals, and p-values.
}
\description{
This function returns a formatted table of generalized linear model results.
}
\examples{
## Basic logistic regression example
n <- 50
x <- rnorm(n)
y <- sample(c(TRUE, FALSE), size=n, replace=TRUE)
fit <- glm(y~x, family=binomial)
glm_table(fit)

## Use se="robust" to get risk ratios from poisson regression
fit <- glm(y~x, family=poisson)
glm_table(fit, se="robust")

## Get a nicely formatted table
glm_table(fit, fmt=TRUE)
}
