% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R, R/summary_measures2.R
\name{n_perc}
\alias{n_perc}
\alias{median_iqr}
\alias{median_range}
\alias{mean_sd}
\alias{n_perc}
\alias{median_iqr}
\alias{median_range}
\alias{mean_sd}
\title{Functions to generate nice-looking summary measures in \code{table_one}}
\usage{
n_perc(x, ag.by, fun_n = n_fmt, fun_prop = prop_fmt, margin = 2)

median_iqr(x, ag.by, fun_median = median_fmt, fun_iqr = iqr_fmt,
  na.rm = TRUE)

median_range(x, ag.by, fun_median = median_fmt, fun_range = range_fmt,
  na.rm = TRUE)

mean_sd(x, ag.by, fun_mean = mean_fmt, fun_sd = sd_fmt, na.rm = TRUE)

n_perc(x, ag.by, fun_n = n_fmt, fun_prop = prop_fmt, margin = 2)

median_iqr(x, ag.by, fun_median = median_fmt, fun_iqr = iqr_fmt,
  na.rm = TRUE)

median_range(x, ag.by, fun_median = median_fmt, fun_range = range_fmt,
  na.rm = TRUE)

mean_sd(x, ag.by, fun_mean = mean_fmt, fun_sd = sd_fmt, na.rm = TRUE)
}
\arguments{
\item{x}{Variable to be summarized}

\item{ag.by}{Variable to aggregate by, usually a character or factor. Optional.}

\item{fun_n, fun_prop, fun_median, fun_iqr, fun_range, fun_mean, fun_sd}{Formatting functions passed as arguments.}

\item{margin}{Integer. 1=row percent, 2=column percent (default).}

\item{na.rm}{Logical. Remove NAs from calculation or return NA if present.}

\item{x}{Variable to be summarized}

\item{ag.by}{Variable to aggregate by, usually a character or factor. Optional.}

\item{fun_n, fun_prop, fun_median, fun_iqr, fun_range, fun_mean, fun_sd}{Formatting functions passed as arguments.}

\item{margin}{Integer. 1=row percent, 2=column percent (default).}

\item{na.rm}{Logical. Remove NAs from calculation or return NA if present.}
}
\description{
Functions to generate nice-looking summary measures in \code{table_one}

Functions to generate nice-looking summary measures in \code{table_one}
}
