% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Produce formatted descriptives tables}
\usage{
table_one(vars, varlabels = vars, data, strata, normal = NULL,
  exact = NULL, fun_n_prc = n_perc, fun_apprx_p = p_cat_apprx,
  fun_exact_p = p_cat_exact, fun_norm = mean_sd, fun_nonnorm = median_iqr,
  fun_norm_p = p_cont_norm, fun_nonnorm_p = p_cont_nonnorm,
  fun_p_fmt = p_fmt, measurelab_nonnormal = "Median [IQR]",
  measurelab_normal = "Mean±SD", measurelab_cat = "n (\%)", sep = " -- ",
  nspaces = 6)
}
\arguments{
\item{vars}{Character vector. The column names of variables to summarize.}

\item{varlables}{Character vector. The labels for variables in the returned table. Should be the same length as \code{var}.}
}
\description{
This function mirrors createTableOne, but allows for more customization.
}
